+++ lib/stat_power.rb
@@ -2,9 +2,12 @@
 
 require_relative "stat_power/version"
 require_relative "stat_power/result"
+require_relative "stat_power/power_result"
 require_relative "stat_power/errors"
 require_relative "stat_power/distributions/normal"
 require_relative "stat_power/solvers/bisection"
+require_relative "stat_power/effect_size/conventional"
+require_relative "stat_power/normal_mean"
 
 module StatPower
 end